pyproject.toml
[project]
name = "exam-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.2",
    "gunicorn>=23.0.0",
    "flask-jwt-extended>=4.7.1",
    "flask-sqlalchemy>=3.1.1",
    "google-genai>=1.61.0",
    "numpy<2",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.9",
    "pillow>=12.1.0",
    "pymupdf>=1.26.7",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.2.2",
    "tenacity>=9.1.2",
    "psycopg[binary]>=3.2.0",
]
